feat: 🎸 use generic implementation for AVL map
